Use global fixtures to reduce the amount of mocking
